Remove assert statement from non-test files

+++ test/auto-updater/test_check_for_updates.py
@@ -47,10 +47,13 @@ def test_update_checker(tmp_path):
 
     # Check for updates
     print(clone_dir)
-    assert update_repository() == True
-    assert update_repository() == False
+    if update_repository() != True:
+        raise AssertionError
+    if update_repository() != False:
+        raise AssertionError
     reset_update_state()
-    assert update_repository() == True
+    if update_repository() != True:
+        raise AssertionError
 
     # Update checking should work with a modified 'master'
     reset_update_state()
@@ -58,8 +61,10 @@ def test_update_checker(tmp_path):
     p.write_text(u'second_file')
     clone_repo.index.add(['second_file'])
     clone_repo.index.commit('modified clone')
-    assert update_repository() == True
-    assert update_repository() == False
+    if update_repository() != True:
+        raise AssertionError
+    if update_repository() != False:
+        raise AssertionError
 
     # Update checking should not work with another branch
     reset_update_state()
@@ -68,11 +73,14 @@ def test_update_checker(tmp_path):
     p.write_text(u'my_branch')
     clone_repo.index.add(['my_branch'])
     clone_repo.index.commit('my_branch')
-    assert update_repository() == False
+    if update_repository() != False:
+        raise AssertionError
 
     # Update checking should not work without a git checkout
     clone_repo.git.checkout('master')
-    assert update_repository() == True
+    if update_repository() != True:
+        raise AssertionError
     reset_update_state()
     rmtree(str((clone_dir / '.git').resolve()))
-    assert update_repository() == False
+    if update_repository() != False:
+        raise AssertionError
